requirements.txt
google-cloud-bigquery
python-dotenv
google-cloud-pubsub